Created an app component to better organize client code
